refactor : test code 분리 (#731)
